requirements.txt
deepface==0.0.79
Flask==3.0.3
Pillow==10.4.0
tensorflow==2.16.1
gunicorn
flask_cors